Bookin appointment implemented

+++ SecureDigitalHealthcare/Controllers/UsersController.cs
@@ -10,16 +10,18 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.Extensions.Hosting;
 using Newtonsoft.Json;
+using EasyHealth.Models;
+using Microsoft.AspNetCore.Authorization;
 using SecureDigitalHealthcare.Models;
 
-namespace SecureDigitalHealthcare.Controllers
+namespace EasyHealth.Controllers
 {
     public class UsersController : Controller
     {
-        private readonly SecureDigitalHealthcareContext _context;
+        private readonly EasyHealthContext _context;
         private readonly IWebHostEnvironment _environment;
 
-        public UsersController(SecureDigitalHealthcareContext context, IWebHostEnvironment environment)
+        public UsersController(EasyHealthContext context, IWebHostEnvironment environment)
         {
             _context = context;
             _environment = environment;
@@ -38,15 +40,10 @@ namespace SecureDigitalHealthcare.Controllers
         }
 
         // GET: Users/Details/5
-        public async Task<IActionResult> Details(string id)
+        public async Task<IActionResult> Details(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
             var user = await _context.Users
-                .FirstOrDefaultAsync(m => m.NationalId == id);
+                .FirstOrDefaultAsync(m => m.Id == id);
             if (user == null)
             {
                 return NotFound();
@@ -139,6 +136,7 @@ namespace SecureDigitalHealthcare.Controllers
             return false;
         }
 
+        [AllowAnonymous]
         public IActionResult GetImage(string imageName)
         {
             var imagePath = Path.Combine(_environment.GetRootProjectPath(), MyHelper.ProfilePicturesFolderName, imageName);
@@ -156,13 +154,8 @@ namespace SecureDigitalHealthcare.Controllers
             }
         }
         // GET: Users/Edit/5
-        public async Task<IActionResult> Edit(string id)
+        public async Task<IActionResult> Edit(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
             var user = await _context.Users.FindAsync(id);
             if (user == null)
             {
@@ -175,9 +168,9 @@ namespace SecureDigitalHealthcare.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ValidateDNTCaptcha(ErrorMessage = "It is invalid")]
-        public async Task<IActionResult> Edit(string id, User user, IFormFile profilePictureInput)
+        public async Task<IActionResult> Edit(int id, User user, IFormFile profilePictureInput)
         {
-            if (id != user.NationalId)
+            if (id != user.Id)
             {
                 return NotFound();
             }
@@ -188,8 +181,11 @@ namespace SecureDigitalHealthcare.Controllers
             {
                 try
                 {
-                    var fileToDelete = _context.Users.AsNoTracking().FirstOrDefault(x => x.NationalId == id)?.ProfileImagePath;
-                    DeleteProfileImage(fileToDelete);
+                    var fileToDelete = _context.Users.AsNoTracking().FirstOrDefault(x => x.Id == id)?.ProfileImagePath;
+                    if (fileToDelete is not null)
+                    {
+                        DeleteProfileImage(fileToDelete);
+                    }
                     SaveProfileImage(profilePictureInput, out string fileName);
 
                     user.ProfileImagePath = fileName;
@@ -202,7 +198,7 @@ namespace SecureDigitalHealthcare.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!UserExists(user.NationalId))
+                    if (!UserExists(user.Id))
                     {
                         return NotFound();
                     }
@@ -217,15 +213,10 @@ namespace SecureDigitalHealthcare.Controllers
         }
 
         // GET: Users/Delete/5
-        public async Task<IActionResult> Delete(string id)
+        public async Task<IActionResult> Delete(int id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
             var user = await _context.Users
-                .FirstOrDefaultAsync(m => m.NationalId == id);
+                .FirstOrDefaultAsync(m => m.Id == id);
             if (user == null)
             {
                 return NotFound();
@@ -242,7 +233,11 @@ namespace SecureDigitalHealthcare.Controllers
             var user = await _context.Users.FindAsync(id);
             if (user != null)
             {
-                DeleteProfileImage(user.ProfileImagePath);
+                if (user.ProfileImagePath is not null)
+                {
+                    DeleteProfileImage(user.ProfileImagePath);
+                }
+
                 _context.Users.Remove(user);
             }
 
@@ -250,9 +245,9 @@ namespace SecureDigitalHealthcare.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool UserExists(string id)
+        private bool UserExists(int id)
         {
-            return _context.Users.Any(e => e.NationalId == id);
+            return _context.Users.Any(e => e.Id == id);
         }
     }
 }
